Refactor : 채팅방 입장, 퇴장 서비스 도입
채팅방 입장, 퇴장을 전담하는 서비스를 도입합니다.
