+++ src/main/java/nbbang/com/nbbang/global/socket/SocketSender.java
@@ -8,7 +8,8 @@ import nbbang.com.nbbang.domain.chat.dto.message.ChatSendResponseDto;
 import nbbang.com.nbbang.domain.member.domain.Member;
 import nbbang.com.nbbang.domain.party.service.PartyService;
 import nbbang.com.nbbang.global.interceptor.CurrentMember;
-import org.springframework.messaging.simp.SimpMessagingTemplate;
+import nbbang.com.nbbang.global.socket.redisPubSub.RedisPublisher;
+import nbbang.com.nbbang.global.socket.redisPubSub.RedisTopicRepository;
 import org.springframework.stereotype.Component;
 
 import java.util.List;
